+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -65,6 +65,7 @@ from rana_qgis_plugin.utils import (
     display_bytes,
     elide_text,
     format_activity_time,
+    get_timestamp_as_numeric_item,
 )
 from rana_qgis_plugin.utils_api import (
     get_frontend_settings,
@@ -255,14 +256,9 @@ class RevisionsView(QWidget):
                 self.revisions_model.setHorizontalHeaderLabels(
                     ["Timestamp", "Event", "Simulation", "Rana Model"]
                 )
-            locel_time = convert_to_local_time(commit_date)
+            commit_item = get_timestamp_as_numeric_item(commit_date)
             if latest:
-                locel_time += " (latest)"
-            commit_item = NumericItem(locel_time)
-            # Use numeric timestamp for sorting
-            commit_item.setData(
-                convert_to_timestamp(commit_date), role=Qt.ItemDataRole.UserRole
-            )
+                commit_item.setText(commit_item.text() + " (latest)")
             # We store the revision object for loading specific revisions in menu_requested.
             if threedi_revision:
                 commit_item.setData((threedi_revision, threedi_schematisation))
@@ -407,6 +403,7 @@ class FileView(QWidget):
         self.selected_file = selected_file
 
     def show_selected_file_details(self, selected_file):
+        tooltips = {}
         self.update_selected_file(selected_file)
         filename = os.path.basename(selected_file["id"].rstrip("/"))
         username = (
@@ -421,6 +418,9 @@ class FileView(QWidget):
         description = descriptor["description"] if descriptor else None
 
         last_modified = convert_to_local_time(selected_file["last_modified"])
+        display_last_modified = format_activity_time(selected_file["last_modified"])
+        if last_modified != display_last_modified:
+            tooltips["Last modified"] = last_modified
         size = (
             display_bytes(selected_file["size"])
             if data_type != "threedi_schematisation"
@@ -432,7 +432,7 @@ class FileView(QWidget):
             ("File type", selected_file["media_type"]),
             ("Data type", SUPPORTED_DATA_TYPES.get(data_type, data_type)),
             ("Added by", username),
-            ("Last modified", last_modified),
+            ("Last modified", display_last_modified),
             ("Description", description),
         ]
         if data_type == "scenario" and meta:
@@ -498,6 +498,8 @@ class FileView(QWidget):
             value_item.setFlags(
                 Qt.ItemFlag.ItemIsEnabled | Qt.ItemFlag.ItemIsSelectable
             )
+            if label in tooltips:
+                value_item.setToolTip(tooltips[label])
             self.file_table_widget.setItem(i, 0, label_item)
             self.file_table_widget.setItem(i, 1, value_item)
         self.file_table_widget.resizeColumnsToContents()
@@ -721,12 +723,7 @@ class FilesBrowser(QWidget):
                 file["size"] if data_type != "threedi_schematisation" else -1,
                 role=Qt.ItemDataRole.UserRole,
             )
-            last_modified = convert_to_local_time(file["last_modified"])
-            last_modified_timestamp = convert_to_timestamp(file["last_modified"])
-            last_modified_item = NumericItem(last_modified)
-            last_modified_item.setData(
-                last_modified_timestamp, role=Qt.ItemDataRole.UserRole
-            )
+            last_modified_item = get_timestamp_as_numeric_item(file["last_modified"])
             # Add items to the model
             self.files_model.appendRow(
                 [name_item, data_type_item, size_item, last_modified_item]
@@ -758,6 +755,7 @@ class ProjectsBrowser(QWidget):
         self.fetch_projects()
         self.populate_projects()
         self.populate_contributors()
+        self.projects_tv.header().setSortIndicator(2, Qt.SortOrder.DescendingOrder)
 
     def set_project_from_id(self, project_id: str):
         for project in self.projects:
@@ -815,14 +813,12 @@ class ProjectsBrowser(QWidget):
     def refresh(self):
         self.current_page = 1
         self.fetch_projects()
-        if not changes:
-            return
         if self.filter_active:
             self.filter_projects()
         else:
             self.populate_projects()
         self.populate_contributors()
-        self.projects_tv.header().setSortIndicator(1, Qt.SortOrder.AscendingOrder)
+        self.projects_tv.header().setSortIndicator(2, Qt.SortOrder.DescendingOrder)
         self.projects_refreshed.emit()
 
     @property
@@ -898,16 +894,7 @@ class ProjectsBrowser(QWidget):
         name_item.setToolTip(project_name)
         name_item.setData(project, role=Qt.ItemDataRole.UserRole)
         # Process last activity time into item
-        last_activity = project["last_activity"]
-        last_activity_timestamp = convert_to_timestamp(last_activity)
-        display_last_activity = format_activity_time(last_activity)
-        last_activity_localtime = convert_to_local_time(last_activity)
-        last_activity_item = NumericItem(display_last_activity)
-        last_activity_item.setData(
-            last_activity_timestamp, role=Qt.ItemDataRole.UserRole
-        )
-        if last_activity_localtime != display_last_activity:
-            last_activity_item.setToolTip(last_activity_localtime)
+        last_activity_item = get_timestamp_as_numeric_item(project["last_activity"])
         # process list of contributors into items
         contributors_item = QStandardItem()
         contributors_data = []